spam_classifier_project/requirements.txt
streamlit
pandas
scikit-learn
pypdf
python-docx
wordcloud
matplotlib